Create separate file for SnailPublisher
